feat: add useCluster
