+++ packages/studio-graph/src/index.tsx
@@ -3,3 +3,4 @@ export { default as ExploreGraph } from './app/explore';
 export { getStyleConfig, getDataMap } from './components/Prepare/utils';
 export * from './components';
 export { useContext } from './hooks/useContext';
+export { default as useCluster } from './hooks/useCluster';
